fix: cleanup remaining Int IDs in Export, Dialogs, and ApprovalHelper
- ClientsImportManager: remove .toInt() on building ID
- PaymentsExportManager: 4x buildingFilter Int?→String?
- ExportDialog: buildingFilter Int?→String?
- PaymentExportDialog: buildingFilter Int?→String?
- ApprovalHelper: targetId Int?→String?

+++ app/src/main/java/com/pronetwork/app/export/ClientsImportManager.kt
@@ -106,7 +106,7 @@ class ClientsImportManager(
                         // إنشاء المبنى الجديد تلقائياً
                         val newBuilding = Building(name = buildingName)
                         val newId = buildingRepository.insert(newBuilding)
-                        building = newBuilding.copy(id = newId.toInt())
+                        building = newBuilding.copy(id = newId)
                         existingBuildings.add(building)
                         newBuildingsCount++
                     }
